Single Script
This is combined script. I have combined the two separate scripts. This script can be run in QGIS python shell. It will ask the user to input various inputs. The user will have to just change the script to put the path of output of fire_frequency.csv because writing it in the directory of script causes os error 30 read only file system as the input directory is read only.